requirements.txt
flask==2.2.3
torch==2.0.0
openai==0.27.4
openai-whisper==20230314
transformers==4.27.4